+++ app/auth/routes.py
@@ -19,7 +19,7 @@ def login():
             User.query.filter_by(username=form.entity.data).first()
         if user is None or not user.verify_password(form.password.data):
             flash('Invalid username or password', 'danger')
-            return redirect(url_for('auth.login'))
+            return redirect(url_for('auth.login'), 401)
         login_user(user, remember=form.remember_me.data)
         next_page = request.args.get('next')
         if not next_page or url_parse(next_page).netloc != '':
