+++ app/templates/navigator.html
@@ -108,7 +108,7 @@
                     </g>
                     </g>
                   </svg>  
-                  <span class="items__count">{{ current_user.cart.cart_products|length }}</span>
+                  <span class="items__count">{{ current_user.cart.cart_products.all()|length }}</span>
                 </a>
               </li>
               <!-- wishlist button -->
@@ -117,7 +117,7 @@
                   <svg xmlns="http://www.w3.org/2000/svg" width="18.541" height="15.557" viewBox="0 0 18.541 15.557">
                     <path  d="M71.775,135.51a5.153,5.153,0,0,1,1.267-1.524,4.986,4.986,0,0,1,6.584.358,4.728,4.728,0,0,1,1.174,4.914,10.458,10.458,0,0,1-2.132,3.808,22.591,22.591,0,0,1-5.4,4.558c-.445.282-.9.549-1.356.812a.306.306,0,0,1-.254.013,25.491,25.491,0,0,1-6.279-4.8,11.648,11.648,0,0,1-2.52-4.009,4.957,4.957,0,0,1,.028-3.787,4.629,4.629,0,0,1,3.744-2.863,4.782,4.782,0,0,1,5.086,2.447c.013.019.025.034.057.076Z" transform="translate(-62.498 -132.915)" fill="currentColor"/>
                   </svg>
-                  <span class="items__count">{{ current_user.cart.cart_products|length }}</span> 
+                  <span class="items__count">{{ current_user.cart.cart_products.all()|length }}</span> 
                 </a>
               </li>
             {% else %}
@@ -457,7 +457,7 @@
               </svg> 
             </span>
             <span class="offcanvas__stikcy--toolbar__label">Cart</span>
-            <span class="items__count">{{ current_user.cart.cart_products|length }}</span> 
+            <span class="items__count">{{ current_user.cart.cart_products.all()|length }}</span> 
           </a>
         </li>
         <!-- wishlist bottom scren small -->
@@ -502,7 +502,7 @@
         </div>
         <p class="minicart__header--desc">{{ config['TITLE'] }}</p>
       </div>
-      {% for cart_product in current_user.cart.cart_products %}
+      {% for cart_product in current_user.cart.cart_products.all() %}
         {% with product=cart_product.product %}
           <div class="minicart__product" id="minicartProduct{{ product.id }}">
             <div class="minicart__product--items d-flex">
@@ -542,7 +542,7 @@
         </div>
       </div>
       <div class="minicart__button d-flex justify-content-center">
-        <a class="btn minicart__button--link" href="{{ url_for('api.cart') }}">View cart</a>
+        <a class="btn minicart__button--link" href="{{ url_for('main.cart') }}">View cart</a>
         <a class="btn minicart__button--link" href="{{ url_for('main.checkout') }}">Checkout</a>
       </div>
     </div>
